Start project refactoring
