--- src/__version__.py
@@ -1,1 +1,0 @@
-version = "1.4.1"
